input directive moved inside directives folder
