+++ src/directives/input/input.directive.spec.ts
@@ -4,7 +4,7 @@ import {
   TestBed
 } from "@angular/core/testing";
 import { By } from "@angular/platform-browser";
-import { IgxLabelModule } from "../label/label.directive";
+import { IgxLabelModule } from "../../label/label.directive";
 import { IgxInputModule } from "./input.directive";
 
 describe("IgxInput", () => {
