+++ src/main.ts
@@ -21,12 +21,14 @@ export * from "./grid/";
 export * from "./navigation-drawer/navigation-drawer.component";
 export * from "./icon/icon.component";
 export * from "./directives/button/button.directive";
+export * from "./directives/input/input.directive";
+export * from "./directives/filter.directive";
+export * from "./directives/ripple.directive";
 export * from "./buttonGroup/buttonGroup.component";
 export * from "./list/list.component";
 export * from "./tabbar/tabbar.component";
 export * from "./calendar/calendar.component";
 export * from "./carousel/carousel.component";
-export * from "./input/input.directive";
 export * from "./checkbox/checkbox.component";
 export * from "./radio/radio.component";
 export * from "./label/label.directive";
@@ -36,8 +38,6 @@ export * from "./layout/layout.directive";
 export * from "./dialog/dialog.component";
 export * from "./progressbar/progressbar.component";
 export * from "./navbar/navbar.component";
-export * from "./directives/filter.directive";
-export * from "./directives/ripple.directive";
 export * from "./badge/badge.component";
 export * from "./snackbar/snackbar.component";
 export * from "./card/card.component";
